1. Change to use BigDecimal for price related code 2. Move main class to its own java file

+++ src/Beauty.java
@@ -1,5 +1,7 @@
+import java.math.BigDecimal;
+
 public class Beauty extends Product{
     public Beauty (String name, int price, int weight) {
-        super(name, price, weight);
+        super(name, BigDecimal.valueOf(price), weight);
     }
 }
